requirements.txt
click==8.1.0
PyGithub==2.1.1
tenacity==8.1.0